+++ src/foundations.jl
@@ -56,8 +56,7 @@ for sym in (Symbol("@callable"),
             :axpby_yields_axmy,
             :axpby_yields_axpby,
             :multiplier_type,
-            :multiplier_floatingpoint_type,
-            :promote_multiplier)
+            :convert_multiplier)
     @eval begin
         import ..LazyAlgebra: $sym
         export $sym